Refactor codebase, add admin panel, and update images
This commit addresses the user's request to optimize the codebase, add new images, and implement an admin panel.

The key changes are:

- **Code Refactoring:** The monolithic `App.tsx` file has been broken down into smaller, reusable components, improving maintainability and scalability. The components are now located in the `src/components` directory.
- **Data Externalization:** The product and hero image data has been moved from being hardcoded in the `App.tsx` file to a separate `src/data.json` file. This makes the data easier to manage and update.
- **Image Updates:** The product images have been updated with new images from Unsplash.
- **Admin Panel:** A new admin panel has been added, accessible at the `/admin` route. The admin panel is protected by a simple hardcoded password. It allows the admin to:
    - View all products.
    - Edit the price of each product.
    - Update the image URL for each product.
- **Save Functionality:** The admin panel includes a "Save Changes" button that generates and downloads a new `data.json` file with the updated product information. The user can then manually replace the old file with the new one.
- **Build Configuration:** A `.gitignore` file has been added to exclude the `node_modules` directory from version control.
