+++ src/components/PageShell.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { motion } from "framer-motion";
+
+export function PageShell({ children }: { children: React.ReactNode }) {
+  return (
+    <motion.div
+      initial={{ opacity: 0, y: 8 }}
+      animate={{ opacity: 1, y: 0 }}
+      exit={{ opacity: 0, y: -8 }}
+      transition={{ duration: 0.35, ease: "easeOut" }}
+      className="relative"
+    >
+      {children}
+    </motion.div>
+  );
+}
