+++ src/types.ts
@@ -1,0 +1,2 @@
+export const pages = ["home", "about", "collections", "contact", "admin"] as const;
+export type Page = (typeof pages)[number];
